+++ widgets/FullCalendar.php
@@ -4,11 +4,12 @@ namespace humhub\modules\calendar\widgets;
 
 use humhub\modules\calendar\assets\CalendarAsset;
 use humhub\modules\calendar\helpers\CalendarUtils;
+use humhub\modules\calendar\helpers\Url;
+use humhub\modules\calendar\models\participation\FullCalendarSettings;
 use humhub\modules\calendar\permissions\CreateEntry;
 use humhub\modules\content\components\ContentContainerActiveRecord;
 use humhub\widgets\JsWidget;
 use Yii;
-use humhub\modules\calendar\helpers\Url;
 
 /**
  * Description of FullCalendarWidget
@@ -39,33 +40,36 @@ class FullCalendar extends JsWidget
     {
         CalendarAsset::register($this->getView());
 
-        if(Yii::$app->user->isGuest) {
+        if (Yii::$app->user->isGuest) {
             $this->canWrite = false;
             $this->enabled = false;
             parent::init();
             return;
         }
 
-        if(!$this->contentContainer) {
+        if (!$this->contentContainer) {
             $this->contentContainer = Yii::$app->user->getIdentity();
             $this->isGlobal = true;
         }
 
         // Used by the global calendar if the module is not enabled for the given user.
-        if($this->contentContainer && !$this->contentContainer->moduleManager->isEnabled('calendar')) {
+        if ($this->contentContainer && !$this->contentContainer->moduleManager->isEnabled('calendar')) {
             $this->enabled = false;
         }
 
-        if($this->contentContainer) {
+        if ($this->contentContainer) {
             $this->editUrl = Url::toFullCalendarEdit($this->contentContainer);
         }
 
         parent::init();
     }
-    
+
     public function getData()
     {
+        $defaultSettings = new FullCalendarSettings(['contentContainer' => $this->contentContainer]);
+
         return [
+            'default-view' => $defaultSettings->viewMode,
             'height' => $this->height,
             'load-url' => $this->loadUrl,
             'aspect-ratio' => $this->aspectRatio,
@@ -80,7 +84,7 @@ class FullCalendar extends JsWidget
             'select-helper' => $this->canWrite,
             'selectors' => $this->selectors,
             'filters' => $this->filters,
-            'time-zone' =>  CalendarUtils::getUserTimeZone(true),
+            'time-zone' => CalendarUtils::getUserTimeZone(true),
             'locale' => $this->translateLocale(Yii::$app->formatter->locale),
         ];
     }
@@ -93,7 +97,7 @@ class FullCalendar extends JsWidget
     private function translateLocale($locale)
     {
         $locale = str_replace('_', '-', $locale);
-        if(array_key_exists($locale, self::LOCALE_MAPPING)) {
+        if (array_key_exists($locale, self::LOCALE_MAPPING)) {
             $locale = self::LOCALE_MAPPING[$locale];
         }
 
@@ -102,9 +106,9 @@ class FullCalendar extends JsWidget
 
     private function canCreate()
     {
-        if($this->contentContainer && !Yii::$app->user->isGuest) {
+        if ($this->contentContainer && !Yii::$app->user->isGuest) {
             return $this->contentContainer->can(CreateEntry::class);
-        } else if(!Yii::$app->user->isGuest) {
+        } else if (!Yii::$app->user->isGuest) {
             return Yii::$app->user->getIdentity()->isCurrentUser();
         }